fixes import paths

+++ src/Galadriel/Gal.sol
@@ -3,20 +3,19 @@ pragma solidity ^0.8.9;
 
 // Uncomment this line to use console.log
 // import "hardhat/console.sol";
-import "./interfaces/IOracle.sol";
+import "./../interfaces/IOracle.sol";
 
 contract OpenAiChatGptVision {
-
     struct ChatRun {
         address owner;
         IOracle.Message[] messages;
-        uint messagesCount;
+        uint256 messagesCount;
     }
 
-    mapping(uint => ChatRun) public chatRuns;
-    uint private chatRunsCount;
+    mapping(uint256 => ChatRun) public chatRuns;
+    uint256 private chatRunsCount;
 
-    event ChatCreated(address indexed owner, uint indexed chatId);
+    event ChatCreated(address indexed owner, uint256 indexed chatId);
 
     address private owner;
     address public oracleAddress;
@@ -31,19 +30,19 @@ contract OpenAiChatGptVision {
         chatRunsCount = 0;
 
         config = IOracle.OpenAiRequest({
-            model : "gpt-4-turbo",
-            frequencyPenalty : 21, // > 20 for null
-            logitBias : "", // empty str for null
-            maxTokens : 1000, // 0 for null
-            presencePenalty : 21, // > 20 for null
-            responseFormat : "{\"type\":\"text\"}",
-            seed : 0, // null
-            stop : "", // null
-            temperature : 10, // Example temperature (scaled up, 10 means 1.0), > 20 means null
-            topP : 101, // Percentage 0-100, > 100 means null
-            tools : "",
-            toolChoice : "", // "none" or "auto"
-            user : "" // null
+            model: "gpt-4-turbo",
+            frequencyPenalty: 21, // > 20 for null
+            logitBias: "", // empty str for null
+            maxTokens: 1000, // 0 for null
+            presencePenalty: 21, // > 20 for null
+            responseFormat: "{\"type\":\"text\"}",
+            seed: 0, // null
+            stop: "", // null
+            temperature: 10, // Example temperature (scaled up, 10 means 1.0), > 20 means null
+            topP: 101, // Percentage 0-100, > 100 means null
+            tools: "",
+            toolChoice: "", // "none" or "auto"
+            user: "" // null
         });
     }
 
@@ -62,28 +61,20 @@ contract OpenAiChatGptVision {
         emit OracleAddressUpdated(newOracleAddress);
     }
 
-    function startChat(string memory message, string[] memory imageUrls) public returns (uint i) {
+    function startChat(string memory message, string[] memory imageUrls) public returns (uint256 i) {
         ChatRun storage run = chatRuns[chatRunsCount];
 
         run.owner = msg.sender;
-        IOracle.Message memory newMessage = IOracle.Message({
-            role: "user",
-            content: new IOracle.Content[](imageUrls.length + 1)
-        });
-        newMessage.content[0] = IOracle.Content({
-            contentType: "text",
-            value: message
-        });
-        for (uint u = 0; u < imageUrls.length; u++) {
-            newMessage.content[u + 1] = IOracle.Content({
-                contentType: "image_url",
-                value: imageUrls[u]
-            });
+        IOracle.Message memory newMessage =
+            IOracle.Message({role: "user", content: new IOracle.Content[](imageUrls.length + 1)});
+        newMessage.content[0] = IOracle.Content({contentType: "text", value: message});
+        for (uint256 u = 0; u < imageUrls.length; u++) {
+            newMessage.content[u + 1] = IOracle.Content({contentType: "image_url", value: imageUrls[u]});
         }
         run.messages.push(newMessage);
         run.messagesCount = 1;
 
-        uint currentId = chatRunsCount;
+        uint256 currentId = chatRunsCount;
         chatRunsCount = chatRunsCount + 1;
 
         IOracle(oracleAddress).createOpenAiLlmCall(currentId, config);
@@ -93,7 +84,7 @@ contract OpenAiChatGptVision {
     }
 
     function onOracleOpenAiLlmResponse(
-        uint runId,
+        uint256 runId,
         IOracle.OpenAiResponse memory response,
         string memory errorMessage
     ) public onlyOracle {
@@ -104,19 +95,13 @@ contract OpenAiChatGptVision {
         );
 
         if (!compareStrings(errorMessage, "")) {
-            IOracle.Message memory newMessage = IOracle.Message({
-                role: "assistant",
-                content: new IOracle.Content[](1)
-            });
+            IOracle.Message memory newMessage = IOracle.Message({role: "assistant", content: new IOracle.Content[](1)});
             newMessage.content[0].contentType = "text";
             newMessage.content[0].value = errorMessage;
             run.messages.push(newMessage);
             run.messagesCount++;
         } else {
-            IOracle.Message memory newMessage = IOracle.Message({
-                role: "assistant",
-                content: new IOracle.Content[](1)
-            });
+            IOracle.Message memory newMessage = IOracle.Message({role: "assistant", content: new IOracle.Content[](1)});
             newMessage.content[0].contentType = "text";
             newMessage.content[0].value = response.content;
             run.messages.push(newMessage);
@@ -124,20 +109,16 @@ contract OpenAiChatGptVision {
         }
     }
 
-    function addMessage(string memory message, uint runId) public {
+    function addMessage(string memory message, uint256 runId) public {
         ChatRun storage run = chatRuns[runId];
         require(
-            keccak256(abi.encodePacked(run.messages[run.messagesCount - 1].role)) == keccak256(abi.encodePacked("assistant")),
+            keccak256(abi.encodePacked(run.messages[run.messagesCount - 1].role))
+                == keccak256(abi.encodePacked("assistant")),
             "No response to previous message"
         );
-        require(
-            run.owner == msg.sender, "Only chat owner can add messages"
-        );
+        require(run.owner == msg.sender, "Only chat owner can add messages");
 
-         IOracle.Message memory newMessage = IOracle.Message({
-            role: "user",
-            content: new IOracle.Content[](1)
-        });
+        IOracle.Message memory newMessage = IOracle.Message({role: "user", content: new IOracle.Content[](1)});
         newMessage.content[0].contentType = "text";
         newMessage.content[0].value = message;
         run.messages.push(newMessage);
@@ -146,7 +127,7 @@ contract OpenAiChatGptVision {
         IOracle(oracleAddress).createOpenAiLlmCall(runId, config);
     }
 
-    function getMessageHistory(uint chatId) public view returns (IOracle.Message[] memory) {
+    function getMessageHistory(uint256 chatId) public view returns (IOracle.Message[] memory) {
         return chatRuns[chatId].messages;
     }
